Split off proxy to individual classes (proxy vs injected-proxy)
