+++ apps/extension/src/provider/Proxy.ts
@@ -1,41 +1,9 @@
 import { Anoma } from "./Anoma";
-import { Result } from "router/types";
-
-export type ProxyMethod = keyof Omit<Anoma, "version">;
-
-export enum ProxyRequestTypes {
-  Request = "anoma-proxy-request",
-  Response = "anoma-proxy-request-response",
-}
-
-// TODO: Look into better generic typing for this:
-// eslint-disable-next-line
-export interface ProxyRequest<T = any> {
-  type: ProxyRequestTypes;
-  id: string;
-  method: ProxyMethod;
-  args: T;
-}
-
-export interface ProxyRequestResponse {
-  type: ProxyRequestTypes.Response;
-  id: string;
-  result: Result | undefined;
-}
+import { ProxyRequest, ProxyRequestResponse, ProxyRequestTypes } from "./types";
 
 export class Proxy {
-  static startProxy(
-    anoma: Anoma,
-    eventListener: {
-      addMessageListener: (fn: (e: MessageEvent<ProxyRequest>) => void) => void;
-      postMessage: (message: ProxyRequestResponse) => void;
-    } = {
-      addMessageListener: (fn) => window.addEventListener("message", fn),
-      postMessage: (message) =>
-        window.postMessage(message, window.location.origin),
-    }
-  ): void {
-    eventListener.addMessageListener(async (e) => {
+  static startProxy(anoma: Anoma): void {
+    Proxy.addMessageListener(async (e) => {
       const message = e.data;
 
       if (!message || message.type !== ProxyRequestTypes.Request) {
@@ -58,7 +26,7 @@ export class Proxy {
           },
         };
 
-        eventListener.postMessage(proxyResponse);
+        Proxy.postMessage(proxyResponse);
       } catch (e: unknown) {
         const proxyResponse: ProxyRequestResponse = {
           type: ProxyRequestTypes.Response,
@@ -68,8 +36,16 @@ export class Proxy {
           },
         };
 
-        eventListener.postMessage(proxyResponse);
+        Proxy.postMessage(proxyResponse);
       }
     });
+  }
+
+  static addMessageListener(fn: (e: MessageEvent<ProxyRequest>) => void): void {
+    window.addEventListener("message", fn);
+  }
+
+  static postMessage(message: ProxyRequestResponse): void {
+    window.postMessage(message, window.location.origin);
   }
 }
