+++ apps/extension/src/provider/index.ts
@@ -1,3 +1,5 @@
 export * from "./Anoma";
 export * from "./InjectedAnoma";
+export * from "./InjectedProxy";
 export * from "./Proxy";
+export * from "./types";
